Fix invalid cast from long to int, causing a compare operation to always return FALSE. This resulted in the main loop being indefinitely stuck into write() into a file descriptor.

+++ belle-sip/src/belle_sip_loop.c
@@ -259,7 +259,7 @@ struct belle_sip_main_loop{
 	bctbx_mutex_t timer_sources_mutex;
 #ifndef _WIN32
 	int control_fds[2];
-	unsigned int thread_id;
+	unsigned long thread_id;
 #endif
 };
